+++ packages/draw/CHANGELOG.md
@@ -42,7 +42,7 @@
 -   [#743](https://github.com/worktile/plait/pull/743) [`c8b3e730`](https://github.com/worktile/plait/commit/c8b3e730b6002df0d4226a7901c9718f6d9780ed) Thanks [@huanhuanwa](https://github.com/huanhuanwa)! - fix get vector error when resize to other side
 
 *   [#744](https://github.com/worktile/plait/pull/744) [`712ccb12`](https://github.com/worktile/plait/commit/712ccb12a6113c5b59eb48b022732824ff529044) Thanks [@huanhuanwa](https://github.com/huanhuanwa)! - optimize the code for drawing equal lines
-    add getDirectionFactorByVectorComponent
+    add getDirectionFactorByDirectionComponent
 
 -   [#737](https://github.com/worktile/plait/pull/737) [`91d1685c`](https://github.com/worktile/plait/commit/91d1685c62c4ac4b7717da315e3a32a0838f1174) Thanks [@pubuzhixing8](https://github.com/pubuzhixing8)! - isHorizontalSegment renamed to isOverHorizontal
     isVerticalSegment renamed to isOverVertical